Update and expand default FGD
Added `func_geo` SolidEntity to default FGD. This acts as an alternative to using worldspawn, and is a highly recommended entity by many developers.

Updated the default FGD to include `func_geo`.

Also changed `worldspawn` and `func_detail` to have a `collision_mask` value of 0, as they are static bodies and don't require a collision mask.