Fix: Runtime Building in Exported Projects Using Imported Map Resources
When exporting projects, Godot will convert scenes and resources into binary formats, then use a reference `.import` or `.remap` file to point to the binary resource. QuakeMapFiles would be converted to `res://path/file.map.import`. FuncGodot (and Qodot before it) was not designed to handle this behavior previously, as it was not well understood behavior by the team until recently. Additionally, the QuakeMapFile resource was a bit superfluous until this fix, as it only saved revision numbers since FuncGodot was relying on the existence of a the raw map file itself yet Godot only saves the resource settings the importer provides to it.

This commit modifies multiple internal processes without affecting work flows.

Part of the fix also adds the map file data to the QuakeMapFile resource, so that it gets baked into the resource when the project is exported.

Additional checks for `.import` for map files and `.remap` for materials should alleviate most runtime building issues.

Currently unwrapping UV2s in an exported project at runtime does not yet work. Users are unable to bake LightmapGI at runtime though, so this is acceptable for now.