Removed Func and Trigger Brush Tags from TrenchBroom Config Defaults
Uncertain when these were added as defaults to the TrenchBroomGameConfig resource. Removed as unlike Clip and Skip, Func and Trigger are not necessary to the build process nor are they universal naming conventions for entities. They are still provided as optional brush tags that can be added to a Game Config resource.