Add support for TrenchBroom GameConfig Version 9
TrenchBroom Game Config V9 is coming with the latest TrenchBroom 2024.2. The only change between V8 and V9 is the term "texture" to "material". As such, instead of making a nearly identical entry I opted to just have a small text replacement for backwards compatibility.

Using the "Latest" setting on the Game Config Version property will still format for Version 8 currently. This will be changed when 2024.2 gets an official stable release, as FuncGodot will always default to the latest versions of supported tools.

All current config settings are currently safe.