Add scale factor property and inverse scale factor setter
Also changes Map Settings to be a `@tool` script, so that `scale_factor` can be updated when a user updates `inverse_scale_factor`.

Currently func_godot does a large number of division problems during the build loops, for the purposes of scaling all object and vertex positions. This is part of an effort to swap division ops for multiplication ops.