Fix Shallow Face Angles Not Being Built (#73)
`CMP_EPSILON` was not small enough to allow extremely shallow face plane normals to be returned, causing them to be skipped in the build process despite returning a valid cross product.

Swapping the return statements and switching the conditional to check for valid cross products (anything greater than zero) rather than invalid ones fixes this issue for *most* shallow brushes made in a map editor. Certain extreme cases are still not caught but that may require a rework of GeoGenerator rather than a simple bug fix.