Add fallback Map Settings path to FuncGodotMap and Project Settings
Attempting to load a value from custom Project Settings returns null when the value is not set to a custom value for some reason. This wasn't noticeable in the Editor but broke when attempting Runtime building.

This hot fix adds the default map settings path as an initial value for the Project Settings and provides it as a fallback value for the FuncGodotMap's map_settings. Runtime building should work again now as a result. Tested against the FuncGodot Runtime Building project.