Fix default value for target property types
Oversight on the target_source / target_destination fix. While the property types were fixed to generate in an FGD correctly, the default values (values used when not found in the map file) would yield `null` for target_source and an empty `NodePath()` for target_destination. They are now correctly parsed as empty strings.