Allow More Flexible Formatting in Entity Descriptions (#85)
Previously entity descriptions were explicitly limited to no special or escape characters. There seemed to be an assumption that the FGD format may not have supported escape characters. Additionally there was an unnecessary replacement of the `\n` escape character as it was immediately followed up by `strip_edges()` which removes `\n` anyway. There was also a superfluous check for the `@BaseClass` prefix value, which was already checked for in a previous conditional.

This PR changes the way the description is processed to only replace the one unsupported escape character: `\"` (descriptions are encapsulated by quote marks, and the FGD format doesn't actually support escape characters, Godot converts them when writing to the FGD file). All other escape characters and symbols appear to be supported by the FGD format, and so are allowed by FuncGodot.