Scale Model Point Entities when Scale Expression set
The _scale expression_ setting can be used to allow a scaling key to
be applied to entities in TrenchBroom, but the map built by func_godot
does not reflect the size as shown in the map editor.

This change scales the generated node if a scale expression is defined
and the entity's properties include a float value for that key. The
final scale is affected by the map's inverse scale factor.