Add Apply Scale On Map Build Option to Point Entities
Added optional `apply_scale_on_map_build` option to the `FuncGodotFGDPointClass` definition. If an entity has the `scale` property on build, it will check whether it's a *float*, *Vector3*, or *Vector2* and multiplies it against the top level node's current scale.

Plugin version also updated to 2025.1 in preparation for next Stable Release.