Prevent Texture Loader from Generating Origin Material
Texture Loader does not create materials for Skip or Clip textures, as the mesh faces using those textures are removed on build. Origin brushes have their faces removed as well, but the texture is still generated.

This PR changes that by adding an exception for the Origin texture alongside Skip and Clip as intended.