FuncGodot Core Rewrite Update
This is a major refactor of all the core classes in FuncGodot, along with some minor interfacing updates. There are a few breaking changes:[br]
- FuncGodotMap's `verify_and_build()` method is now `build()`.
- FuncGodotMap's `unwrap_uv2()` has been removed. In its place is a build flag toggling UV2 unwrapping.
- FuncGodotMapSettings's `use_trenchbroom_groups_hierarchy` property is now `use_groups_hierarchy`.
- FuncGodotLocalConfig's `game_path_models_folder` is now a Project Setting called `func_godot/model_point_class_save_path`.
- FuncGodotLocalConfig's `default_inverse_scale` is now a Project Setting called `func_godot/default_inverse_scale`.

Everything else should remain compatible.[br]

Smooth shading is not yet working correctly. It's still fairly complicated to get performant.[br]

New core classes have replaced the old ones with a focus on simplification and taking advantage of Godot's built-in types and methods. It should be much easier to maintain and modify going forward.

Every script should be fully commented to take advantage of Godot's auto-documentation features.