Fix Clip Texture Bug and Cleanup
When a solid entity had a face textured with the clip texture, func_godot would attempt to generate the mesh surface using the clip material that couldn't be found as it is never added to the `texture_materials` Dictionary generated with FuncGodotUtil. Additionally, since the surface using clip is never generated, this meant that all of the texture offsets were thrown off when using the clip texture on a brush entity.

The fix just involves removing the clip texture entry from the textures array that the mesh generation methods look from, fixing the offset issue and the invalid material lookup issue.

Additionally, many small cleanup changes were made to better align with the GDScript style guide and the rest of the func_godot scripts.