Add Vertex Merge Property Feature on Per Entity Basis
The `vertex_snap_epsilon` property was added to Map Settings in a prior update as a stop gap measure: https://github.com/func-godot/func_godot_plugin/commit/e38ce937ccf854e69c9d5008ca0044a017ed823b

This was decided to be a bit too blunt of a fix and has been changed to a similar design to the `_phong` property.

Map Settings gains a new property: `vertex_merge_distance_property`, which determines what key value pair is used for snapping distance on a solid entity.

A new base class, VertexMergeDistance, has been added to the default entities. This class comes guaranteed with FuncGodot and can be reused freely in projects.

The aim of the property is to reduce seam gaps between polygons (also referred to as "sparkles" or "ants"). Some brush work may require different values to reduce the effect, some may not need it at all; this base class allows users to make that judgment on a per entity basis.