Merge face vertices when doing rounding
This is an attempt to help with situations when floating point
rounding issues cause too many vertices to be generated for a
winding. Usually this just creates tiny triangles that are effectively
invisible, but when used in combination with the 'vertex merge distance'
option, it was causing faces to be visually flipped, in a way I don't
quite understand (since the rounding doesn't change the vertex order).

This hopefully eliminates the problem at the source by merging adjacent
vertices that would become equal once rounding is applied. It also does the
rounding at the same time.