Bake Scale Factor Into Brush Planes At Parser Stage
The introduction of the hyperplane build algorithm exposed an issue with func_godot's floating point errors due to the large numbers inherent in Quake unit measurements. In order to reduce the effect of floating point issues, Quake to Godot unit scaling now occurs at the parsing stage for geometry.

Additionally, `hyperplane_size` has been added to the FuncGodotMap node properties, allowing further control over the build process. Adjusting the hyperplane size may reduce the effect of gaps in the seams between polygons, but be aware that the hyperplane must be able to cover the area of your entire map.