Skip attempting to auto-apply 'scale' property
When auto applying both matching properties by name as well
as auto applying the 'scale' property, an error is printed
when scale is not a Vector3. Unlike the other magic properties
(e.g. mangle/angle/origin) the Node3D name in the Godot node is
the same, so this can't be worked around.

func_godot supports float, string, or Vector types in the
scale property, and it has already been applied in the
`generate_point_entity_node` function, so the easiest way is to
simply skip this property when auto-applying properties of
the same name.