Add Cull Interior Faces Option to Entities (#182)
This PR adds a new entity build property to remove interior faces during build. Interior faces are defined as one or more of the following:

* Faces where all vertices in them are shared with another face (Implemented in this PR). This also handles cases where a triangle prism is directly above a cuboid and they share verts(the cuboid does not lose any faces, the prism loses its triangular face).
* Faces that are completely contained by another the face

To facilitate this, a new base class has been created and added to all of the default entities that adds a `_cull_interior_faces` property, defaulting to `0`.

This only works on a per entity basis and will not cull faces across entities. This also only works between 2 faces - a face will not be culled if its vertices span across 2 brushes.

Large brush entities will create a hit in build performance. Attempts to optimize it should be made in the future, but it works for now as an optional tool.